Add files via upload
This paper presents the design and preliminary development a new model based on blockchain for improving brand loyalty in agribusiness by customer reward. Using facilities of Ethereum Network and solidity programing, we have created an Agri non-fungible token (NFT) according to ERC-721 protocol that is most widely used protocol for creating NFTs on Ethereum. One of the main advantages of ERC-721 is its flexibility, as it allows developers to define custom metadata for each NFT and customize the attributes. The interaction between Ethereum nodes and the user is established via the frontend, which is developed in Nodejs and connects to the Ethereum network. In previous works on loyalty systems, blockchain-based solutions were proposed that resemble the system developed in this study but there are significant differences between our work and other related works such as: the use of NFT and personalization for customers, the adoption of a new version of Solidity, as well as the adoption of the metadata feature for using media to encourage and assist customer brand awareness and repurchase. We tried to provide a safe and transparent way to use NFTs in the way of increasing brand loyalty in agricultural businesses. Due to characters of NFTs as uniqueness and personalization, customers in a loyalty cycle prefer to shop more in order to earn more AgriNFT, causing retailer to sell more items and manufacturers to generate more goods.